packages/dkdc-dl/pyproject.toml
[project]
name = "dkdc-dl"
version = "0.1.1"
description = "dkdc datalake"
readme = "README.md"
authors = [{ name = "Cody", email = "cody@dkdc.dev" }]
requires-python = ">=3.12"
dependencies = [
    "pandas",
    "polars",
    "duckdb==1.3.0",
    "gcsfs>=2025.5.1",
    "ibis-framework[duckdb,postgres]>=10.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
